+++ packages/storage/src/providers/s3/apis/internal/uploadData/multipart/index.ts
@@ -1,0 +1,7 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export {
+	getMultipartUploadHandlers,
+	MultipartUploadDataInput,
+} from './uploadHandlers';