+++ packages/storage/src/providers/s3/index.ts
@@ -48,3 +48,5 @@ export {
 	GetUrlOutput,
 	GetUrlWithPathOutput,
 } from './types/outputs';
+
+export { DEFAULT_PART_SIZE } from './utils/constants';