+++ packages/storage/src/providers/s3/utils/client/s3data/putObject.ts
@@ -19,6 +19,7 @@ import {
 	serializePathnameObjectKey,
 	validateS3RequiredParameter,
 } from '../utils';
+import { validateObjectUrl } from '../../validateObjectUrl';
 
 import { defaultConfig, parseXmlError } from './base';
 import type { PutObjectCommandInput, PutObjectCommandOutput } from './types';
@@ -38,6 +39,7 @@ export type PutObjectInput = Pick<
 	| 'Metadata'
 	| 'Tagging'
 	| 'ChecksumCRC32'
+	| 'ExpectedBucketOwner'
 >;
 
 export type PutObjectOutput = Pick<
@@ -56,12 +58,20 @@ const putObjectSerializer = async (
 			...input,
 			ContentType: input.ContentType ?? 'application/octet-stream',
 		})),
-		...assignStringVariables({ 'content-md5': input.ContentMD5 }),
-		...assignStringVariables({ 'x-amz-checksum-crc32': input.ChecksumCRC32 }),
+		...assignStringVariables({
+			'content-md5': input.ContentMD5,
+			'x-amz-checksum-crc32': input.ChecksumCRC32,
+			'x-amz-expected-bucket-owner': input.ExpectedBucketOwner,
+		}),
 	};
 	const url = new AmplifyUrl(endpoint.url.toString());
 	validateS3RequiredParameter(!!input.Key, 'Key');
 	url.pathname = serializePathnameObjectKey(url, input.Key);
+	validateObjectUrl({
+		bucketName: input.Bucket,
+		key: input.Key,
+		objectURL: url,
+	});
 
 	return {
 		method: 'PUT',